requirements.txt
absl-py==2.3.1
agentic-doc @ git+https://github.com/landing-ai/agentic-doc.git@4791886b1029f98e96a1364dc372ee448284b6d1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.35.1
azure-identity==1.25.1
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
boto3==1.40.51
botocore==1.40.51
brotli==1.1.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cobble==0.1.4
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
datasets==4.2.0
ddgs==9.6.0
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
donut-python==1.0.9
duckduckgo-search==8.1.1
et-xmlfile==2.0.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-documentai==3.6.0
google-cloud-storage==3.4.1
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.41.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
invoke==2.2.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
landingai-ade==0.18.2
langextract==1.0.9
lark==1.3.0
lightning-utilities==0.15.2
lxml==6.0.2
magika==0.6.2
mammoth==1.10.0
markdownify==1.2.0
markitdown==0.1.3
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mistralai==1.9.11
mistune==3.1.4
ml-collections==1.1.0
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
multiprocess==0.70.16
munch==4.0.0
mutagen==1.47.0
narwhals==2.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.7
notebook-shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.1
openai==2.2.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdfminer-six==20250506
pexpect==4.9.0
pillow==11.3.0
pillow-heif==1.1.1
platformdirs==4.4.0
primp==0.15.0
proglog==0.1.12
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.0
pydantic-core==2.41.1
pydantic-settings==2.11.0
pydeck==0.9.1
pydub==0.25.1
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.5
pyparsing==3.2.5
pypdf==5.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-pptx==1.0.2
pytorch-lightning==2.5.5
pytube==15.0.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
s3transfer==0.14.0
safetensors==0.6.2
sconf==0.2.5
seaborn==0.13.2
send2trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
speechrecognition==3.14.3
stack-data==0.6.3
streamlit==1.50.0
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
timm==1.0.20
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
types-jsonschema==4.25.1.20251009
types-python-dateutil==2.9.0.20251008
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.14
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.0.3
yt-dlp==2025.9.26
zss==1.2.0
